+++ Ruge.Win.Test/Ruge.Win.Test/Controls/Text.cs
@@ -13,10 +13,10 @@ namespace Ruge.Win.Test.Controls
 {
     public class Text : TextBlock
     {
-        public Text(string controlid, int width, int height, string text,byte r, byte g, byte b) :base()
+        public Text(string controlid, double width, double height, string text,byte r, byte g, byte b) :base()
         {           
-            SetValue(WidthProperty, (double)width);
-            SetValue(HeightProperty, (double)height);
+            SetValue(WidthProperty, width);
+            SetValue(HeightProperty, height);
             SetValue(TextProperty, text);
             SetValue(FontSizeProperty, new FontSizeConverter().ConvertFromString((height * .75).ToString() + "px"));
             SetValue(PaddingProperty, new Thickness(0, 0, 0, 0));
